+++ src/main/java/mdteam/ait/tardis/data/properties/PropertiesHandler.java
@@ -6,6 +6,7 @@ import mdteam.ait.registry.DesktopRegistry;
 import mdteam.ait.tardis.Tardis;
 import mdteam.ait.tardis.TardisDesktopSchema;
 import mdteam.ait.tardis.data.FuelData;
+import mdteam.ait.tardis.data.TardisCrashData;
 import mdteam.ait.tardis.util.TardisUtil;
 import mdteam.ait.tardis.wrapper.server.manager.ServerTardisManager;
 import net.minecraft.util.Identifier;
@@ -255,6 +256,8 @@ public class PropertiesHandler { // todo move things out of properties
         map.put(MAT_TICKS,0);
         map.put(SPEED,0);
         map.put(IS_CLOAKED, false);
+        map.put(TardisCrashData.TARDIS_RECOVERY_STATE, TardisCrashData.State.NORMAL);
+        map.put(TardisCrashData.TARDIS_REPAIR_TICKS, 0);
 
         unlockAllFreebies(map);