+++ CHANGELOG.md
@@ -8,3 +8,5 @@
 - Update drifting audio file so that it's louder and more up to date
 - Removed depreciation on some functions because I like using them deal with it 😝 - Creativious
 - Changed how crashing works
+- Tardis is now toxic to be in after crashing for a certain amount of time
+- Tardis now won't land correctly and damages itself if you try to fly it while it's unstable